move wasm-opt into new module
