+++ src/lib.rs
@@ -2,6 +2,7 @@ pub mod info;
 pub mod instrumentation;
 pub mod limit_resource;
 pub mod metadata;
+pub mod optimize;
 pub mod shrink;
 pub mod utils;
